refactor: large refactoring of Cli module
