+++ src/commands/mod.rs
@@ -6,7 +6,8 @@ mod replay;
 mod run;
 
 // Re-export only the functions that execute commands
-pub use data::{clean_data, fetch_remote_data, get_data_status, package_data_to_archive};
+pub use data::data_description::DataStatus;
+pub use data::{clean_data, fetch_remote_data, package_data_to_archive};
 
 pub use new::create_kerblam_project;
 pub use other::ignore;
